FREEI-1269: show outbound route Notificaitons feature, if pagingpro is enabled but expired
The pagingpro module hides all of its features once its license expires.  This was unexpected.
The problem was that paging was relying on the existing and enabled pagingpro to handle the feature.
But pagingpro was hiding it because its license  was expired.
This change makes paging handle/display the feature if the pagingpro module is enabled but expired.