FREEI-1241: make Notifications option available when creating a new Outbound Route
Separate PR for FPBX15. Hi Kapil, please let me know if 13 needs a separate PR too.  Copied details from previous PR:
This was an existing behavior for Notifications with pagingpro.
To set the Notifications value, we need to know the route_id that the newly created route ends up with. Doing this in the hookProcess_core function let's us figure that out from the route name. There is some frontend validation that keeps names unique.